[libclang] NFC: Make `CXDiagnosticSetDiagnosticConsumer` reusable
(cherry picked from commit 6982f7f915abecf016c888f30caf4502b48c1c30)
